LayoutTests/imported/w3c/web-platform-tests/intersection-observer/target-in-detached-document-expected.txt


PASS IntersectionObserver in a single document using the implicit root.
FAIL First rAF. assert_equals: No initial notification while in detached document. expected 0 but got 1
